new site in diff folder
a new site was made trying to use jquery. it is found in the site-old folder for now until i can structure it more and add more content to fill in the hole from the current site. to acess the new site, go to the journal entries and click on the site.