microTalk is functional
The tutorial website for microTalk is officially functional. It has a YouTube video embedded, more detailed instructions with reasonings and improved design decisions from Izabella.
Only think left is to add javascript for going back all the way up, or closing the tab. As well as animating when the tab opens, closes, and when switching to a tab.